requirements.txt
argon2-cffi==23.1.0
zstd==1.5.7.0
python-dotenv==1.1.0